requirements.txt
aioredis==2.0.1
anyio==3.5.0
appdirs==1.4.4
asgiref==3.5.2
async-timeout==4.0.2
black==21.6b0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
cleo==1.0.0a4
click==8.1.3
coreapi==2.3.3
coreschema==0.0.4
crashtest==0.3.1
cryptography==37.0.2
defusedxml==0.7.1
Deprecated==1.2.13
Django==3.2.7
django-autoslug==1.9.8
django-countries==7.2.1
django-environ==0.7.0
django-filter==21.1
django-phonenumber-field==5.2.0
django-templated-mail==1.1.1
djangorestframework==3.12.4
djangorestframework-simplejwt==4.8.0
djoser==2.1.0
environ==1.0
fastapi==0.75.0
flake8==3.9.2
fuzzywuzzy==0.18.0
h11==0.13.0
hiredis==2.0.0
httptools==0.4.0
idna==3.3
isort==5.9.3
itypes==1.2.0
jellyfish==0.9.0
Jinja2==3.1.2
MarkupSafe==2.1.1
mccabe==0.6.1
mypy-extensions==0.4.3
oauthlib==3.2.0
packaging==21.3
pathspec==0.9.0
phonenumbers==8.12.33
Pillow==8.3.2
pptree==3.1
psycopg2-binary==2.9.1
py4j==0.10.9.3
pycodestyle==2.7.0
pycparser==2.21
pydantic==1.9.0
pyflakes==2.3.1
PyJWT==2.1.0
pylev==1.4.0
pyparsing==3.0.7
pyspark==3.2.1
python-dotenv==0.19.2
python-ulid==1.1.0
python3-openid==3.2.0
pytz==2022.1
PyYAML==6.0
redis==4.2.0
redis-om==0.0.20
regex==2022.4.24
requests==2.27.1
requests-oauthlib==1.3.1
six==1.16.0
sniffio==1.2.0
social-auth-app-django==4.0.0
social-auth-core==4.2.0
sqlparse==0.4.2
starlette==0.17.1
toml==0.10.2
types-redis==4.1.18
types-six==1.16.12
typing_extensions==4.1.1
uritemplate==4.1.1
urllib3==1.26.9
uvicorn==0.17.6
uvloop==0.16.0
watchgod==0.8.1
websockets==10.2
wrapt==1.14.0
xlrd==1.2.0
xlwt==1.3.0
